+++ packages/core/src/crud/swagger/examples.helper.ts
@@ -10,9 +10,7 @@
  */
 import type { CrudSwaggerSynthExampleFn } from '../../interfaces/crud-swagger-options.interface';
 import { BaseRouteName } from '../../types';
-import { safeRequire } from '../../util';
-
-const swaggerConst = safeRequire('@nestjs/swagger/dist/constants', () => require('@nestjs/swagger/dist/constants'));
+import { swaggerConst } from './swagger-constants';
 
 // cited: node_modules/@nestjs/swagger/dist/constants.js line 15 for
 // DECORATORS.API_MODEL_PROPERTIES_ARRAY, line 14 for DECORATORS.API_MODEL_PROPERTIES.